change addrman::size() to return size_t

+++ src/addrman.h
@@ -393,7 +393,7 @@ public:
     }
 
     // Return the number of (unique) addresses in all tables.
-    int size()
+    size_t size()
     {
         return vRandom.size();
     }
